Enable all I2C buses and UARTs on BeagleBone Black
This was a bit tricky. At first, I wanted to use one single DTB for both
barebox and the Linux kernel, but unfortunately, barebox gets stuck once
I enable uart4 or uart5. That's a bit bummer, so I went back to using
separate device trees for the bootloader and for the Linux kernel.

What remains unsolved is getting all the BBB-specific stuff (or rather
"Black-specific stuff") used without copying them. In the meanwhile,
let's just start on that file, and add our fixes to the bottom of that.

The actual DT fragments were taken from the cape files which are quite
widely available on the web (and marked as (C) CircuitCo). I adopted
them to use vanilla's conventions of pin definitions.

Change-Id: Ie3dc16b527ea55f150d3cbddfa659578f5e2f9e8