HACK: Use a read-only rootfs (almost)
This is supposed to be supported by upstream systemd *and* buildroot by
now, so let's see what breaks. I expect that sysrepo & netopeer won't
like this because of repository paths and what not. It also won't be
possible to change the system's timezone, passwords, network
configuration etc, which is something that we will have to fix before
this works "for real".

Originally I wanted to go and use something like RAUC this week, but it
isn't really trivial. This hack at least works on the CL-ROADMv2. It's
ugly.

Change-Id: I52510509370c2605787a3c33e78884c936ec1717