Enable watchdog on Clearfog
Here's how it works:

- U-Boot configures the watchdog and leaves it running. It will still
reset the watchdog while the CLI is active, though, so interactive
sessions are not really affected by this.

- The Linux kernel simply ignores the watchdog. It won't do anything.

- systemd reconfigures the watchdog and ensures that it's being reset
periodically.

If for some reason systemd cannot configure/reset the watchdog, the
system will eventually reboot because the kernel had not disabled the
watchdog earlier in the boot -- i.e., U-Boot's configuration is still
live.

Change-Id: I1cb7eefe8969139aceba6c0559983240c6dad49e