Pass info about HW model to kernel
The idea here is to decouple "one-time configuration" which is supposed
to happen exactly once during product manufacturing, such as MAC
addresses or, in this case, the information about the HW model, from
user-accessible configuration.

This way, we can have a universal image which applies for all CzechLight
devices that happen to use the Clearfog Base as a controller, regardless
on whether it's a (future) production 1U ROADM line card, or the legacy
CL-ROADMv2 box.

It is expected that during the initial flashing of the system to the
eMMC, the flashing script will store this into U-Boot's environment. The
OS image (a rootfs) will contain systemd unit files for all possible
combinations of CzechLight models along with a
ConditionKernelCommandLine=... stanza to control which of the unit files
to actually start.

Change-Id: I3c1b5d6627538e590b6925cae945e2bbd1a96e34