HACK: Ensure that rootfs is r/w
Well, this is largely a hack. I wanted to ensure that the existing
buildroot options are not modified (I pretty much like the /var-on-tmpfs
which I get for free in R/O rootfs), and at the same time, sysrepo
*needs* to store its data somewhere.

This is intended as a stopgap measure until I'm ready with a setup that
uses a separate data partition.

Change-Id: Ie040ba68266a31e627048b7ea7adcbfdb5e9555f