rauc: Do not require an extra partition
...because we can do better by patching the built-in U-Boot's
configuration. Yup, it's a patch that is going to be quite painful to
maintain, but on the other hand the whole idea here is to *not* touch
the bootloader anyway because that is a risky operation. For that, it's
arguably safer to not rely on yet another piece of configuration (the
boot script) for reliable operation.

Change-Id: Ie559cf07d1dc6e50bc6ff80320509e3933b0af93