Enable the ADT7463 hwmon on relevant boards
The kernel driver says that the chip should have been configured by BIOS
in case it's wired for two-wire fans, and that there's little point in
exporting a userspace configuration interface for this. We don't have
BIOS and we generally do not want to modify U-Boot for stuff that can be
handled outside of the bootloader. Patching kernel for this sounds like
a stupid idea, so the best remaining option is probably to just let the
userspace do this soon enough.

It should be possible to handle additional products (HW different from
CL-ROADMv2) by simply adding extra ConditionKernelCommandLine= stanzas.

In future when we support more mainboards, we should somehow accommodate
for possibly different bus numbers (env file?). Let's just get this done
for now. All changes are in /usr, so we can easily overwrite them via a
RAUC update, anyway.

Change-Id: I95eced7746e4f2048f6a7fdbd8bbcd2e4ddce882