clearfog: RO rootfs, transient RW overlay at /etc
We cannot use systemd's volatile root because that one requires an
initrd which we do not have. Also, my attempts to do this via tmpfiles.d
run into issues with dependencies. Doing this manually is probably the
safest option.

IMHO, there's a lot of assumptions for the writability of /etc. Sysrepo
is just one of them; there's /etc/shadow for managing passwords, and so
on. It's true that many daemons support distincts set of configurations,
one in /etc and another in /run (a tmpfs), but still -- there's always
an issue in *where* to save these bits if we want a R/O rootfs. How do
we mount /etc early enough from within systemd? That appears to be
virtually impossible without an initrd, and I still haven't gotten up on
that idea to not require (and to not have to generate) an initrd...

FIXME: No changes are being saved, for now.

Change-Id: Id9e5110f2876d55b117570a3dfa2ed0520785ff4