clearfog: Store the machine-id in uboot's env
It would be nice to simply save this into /etc/machine-id, but our
overlayfs which is mounted at /etc makes this, er, "tricky" (just take a
look at the etc-overlay.service). We *could* probably do something like
temporarily mounting the rootfs r/w just to update that file at bootup,
and writing it to the other slot during updates from within RAUC, but
that would be yet another special case. How would we handle the initial
flashing, for example?

This approach, while needlessly complicating the bootloader's setup (and
that's a big consideration for me) makes the overlay FS provisioning a
bit easier. We caught that special case of /etc/machine-id being a mount
of tmpfs early enough. The root filesystem remains truly read-only, etc.

Change-Id: I27972ef99b8e19cd44736d062cb5e6e7a1f16bde