Move the fs errors -> panic to the mount options
Previously, we used mkfs options to ensure that the ext4 filesystem
paniced the kernel upon a serious error. With RAUC, we no longer have a
direct control over how the filesystem is created.

This patch ensures that we panic via the mount options instead of mkfs
options. That way we can ensure that this preference is respected no
matter how the FS image was created.

Change-Id: Ie83b299a80960f6ad9e23fdd86552706c71515a4