systemd: Implement the /cfg mount in a generic manner
Our previous method sucked a bit because it relied on hardcoded,
board-specific partition names. It also bypassed systemd's mount
management and used a direct call to mount(1) instead.

So I asked on the systemd mailing list and got an advice to use systemd
generators.

NOTE: Ensure that your build and target trees have been cleared, the
previously-installed cfg-storage.service breaks this one.

Kudos for the sed magic to [1].

[1] https://stackoverflow.com/a/22550640/2245623

Change-Id: I269ce040b2d8ce88083ae0bb667ac12b7a0ee72e