clearfog: Enable MAX14830 quad UART
I had to be very careful when it came to clock naming; simply using
"oscillator" actually mangled the "refclk" on the SoC, which had a nice
side effect of confusing the watchdog timer. Sweet times.

I verified the interrupt lane's mapping by running `gpio-event-mon -n
gpiochip0 -o 22 -r -f` without this patch. It was reacting to a wire
between 3.3V and INT.

Change-Id: I8a0adb3d15586797ffe6d95fda2365df87ba9e36