clearfog: Enable MAX14830 quad UART
It was necessary to use a differently named DTS file, unfortunately, and
that also led to an incompatible change in the bootloader. A full flash
is therefore required. I wasn't able to use the same file name if I want
to avoid duplication by just inlcuding the original DTS board file, and
adding some post hooks for copying just sounded ugly.

Change-Id: I8a0adb3d15586797ffe6d95fda2365df87ba9e36