clearfog: Use a custom DTS file
It was necessary to use a differently named DTS file, unfortunately, and
that also led to an incompatible change in the bootloader. A full flash
is therefore required. I wasn't able to use the same file name if I want
to avoid duplication by just including the original DTS board file, and
adding some post hooks for copying just sounded ugly.

Change-Id: Ia8c11a3f43033a18ed2a8dbae996512e38358367