clearfog: Add a CP2120 SPI-to-I2C bridge
This is partially ugly, but it really works, which is terrific.

Some gotchas/surprises:

- We're using a custom GPIO for SPI CS, and the Orion driver hard-codes
SPI_CS#0 to be used in this purpose. We cannot do that because there's a
NOR flash at spi1.0, hence the spi-gpio-cs.patch.

- Some of the pin muxing is, er, "creative". The RST is Mikrobus' RST,
but the INT pin is at GPIO18. That might once again require additional
pull-ups.

FIXME: clarify which pull-ups are actually needed
FIXME: test which connections actually break PoR

Change-Id: Id83e33c6c1c8ce94a53cf54b0b05fd20034d7b8b
Patch: https://patchwork.ozlabs.org/patch/783087/ .