WIP: clearfog: Add a CP2120 SPI-to-I2C bridge
FIXME: investigate random data corruption, yay

This is partially ugly, but it at least half-way works, which is
terrific.

Some gotchas/surprises:

- The SCLK gets pulled up at reset by the CP2120, which causes the
Clearfog to not boot. Fix this by a buffer?

- We're using a custom GPIO for the SPI CS signal , and the Orion Linux
driver hard-codes SPI_CS#0 to be used in this purpose. We cannot do that
because there's a NOR flash at spi1.0, hence the spi-gpio-cs.patch.

- There should probably be a pull-up at JP18 (IRQ pin of CP2120);
strange that it works without one.

Change-Id: Id83e33c6c1c8ce94a53cf54b0b05fd20034d7b8b
Patch: https://patchwork.ozlabs.org/patch/783087/ .