WIP: clearfog: Add a CP2120 SPI-to-I2C bridge
FIXME: investigate random data corruption, yay

This is partially ugly, but it at least half-way works, which is
terrific.

Some gotchas/surprises:

- The SCLK gets pulled up at reset by the CP2120, which causes the
Clearfog to not boot. Fix this by a buffer?

- We're using a custom GPIO for the SPI CS signal, and the Orion Linux
driver hard-codes SPI_CS#0 to be used in this purpose. We cannot do that
because there's a NOR flash at spi1.0, hence the spi-gpio-cs.patch.

- There's a pull-up on the IRQ line of MAX14830 EV board; our interface
will have to have one as well. J18 on Clearfog (IRQ for CP2120) has a
pull-up on board, so there's nothing to do there.

Change-Id: Id83e33c6c1c8ce94a53cf54b0b05fd20034d7b8b
Patch: https://patchwork.ozlabs.org/patch/783087/ .