WIP: Use bitbang I2C over MAX14830
...for an extra I2C bus which might be eventually needed for our PMBus
purposes. Yes, this one is pretty much crazy; we are only using it as a
proof-of-concept to show that this chip is fast enough to drive SCL at
200kHz.

This needs some not-yet-fully-merged patches to the kernel; they all
have been sent out to the respective mailing lists. Some of them will be
in 4.15, others (like the MAX14830 open-drain mode) will have to wait
until the 4.16's merge window.

Change-Id: If2d69972f31b1db02db300dee75094f5cfecda39