clearfog: USB autoboot on eMMC bring-up
Let's use U-Boot's distro boot configuration for auto-executing the boot
script. This requires a proper name of the file and also a partition
that is marked as bootable.

This only helps on a system that is already jumpered to booting from
UART and where the MMC memory is unpartitioned. If we're recovering from
a broken eMMC flash, this won't really help.

This setup is also not enough to auto-boot from a factory-installed
Marvell-specific U-Boot. Even with the flash disk plugged in, it fails
with a "Bad Linux ARM zImage magic!". We will therefore have to keep
booting from UART for now...

Change-Id: I25caf68e33c47ab326b6c87d3080660575d7ff10