clearfog: Reduce image size so that it fits the eMMC
Yup, a classic error when it comes to sizes:

  mmcblk0: p4 size 4096000 extends beyond EOD, truncated

These are the actual sizes as reported by Linux on a booted system.
There are also two boot partitions, `boot0` and `boot1`.

 - mmcblk0: 15269888
 - boot0: 8192
 - boot1: 8192
 - p1: 4093952
 - p2: 4096000
 - p3: 4093952
 - p4: 2981888

This means that the total usable size on the "8GB eMMC", out-of-box,
is 7.28125 GiB. That's 3727 MB for each slot (two 1MB chunks are
reserved for the U-Boot when on SD-card for symmetry).

Change-Id: I08e35d9a95449dc779883d692c60e129a91a045b