clearfog: Do not mess with the SPI direct mode
I re-read the datasheet, and it seems that the direct mode is really
meant to transactions which somehow target SPI flash. The protocol talks
about addresses and headers and dummy bytes which exactly correspond to
SPI NOR flash IO protocol. It is possible to turn these off and transfer
raw data, though.

When combined with a CS GPIO, this thing breaks horribly. I've seen my
logic analyzer saying that the CS pin went up in the middle of a second
byte, etc. That wasn't fun.

Going further, the Linux kernel commit b3c195b3a75b which added the
direct-mode feature explicitly talks about writes being tested, and
reads being supported, but untested. That's weird because I just cannot
see how that piece of code implements any reads. It seems that support
for reads actually got removed in a v4 posting of this patch, but the
commit message was never updated. Yay.

This reverts commit 8fd3c5d035fba6e4294f6ae57b974477bf97ef94.

Change-Id: Ic3e0939fcafcc94f15eb142ad53a89b66dff2714
Link: https://patchwork.kernel.org/patch/9124561/