clearfog: SPI: Add a hack for reset during boot
We have three external chips whose "lifetime" does not necessarily match
the boot-up state of the main SoC. We also don't have access to the
power-up-reset lanes of the main SoC. Still, it would be nice to
automatically reset both peripherals (one MAX14830 and two MCP23S17)
during bootup so that they are in a "known good" state.

This patch is a hack because there's no dependency saying "create the
hog at first" and only after that "probe the drivers".

FIXME: Check the `regulator-gpio`.

Change-Id: I89b2cba2e912f7aefff9b08dbc9d759cdaa36954