Lumentum OCM: setup the SPI bus
Too bad that we're going to use the J18 header as well. A possible
alternative could be something like a NOR gate at the CS1 and CS2 pins
which triggers when both of them are at logic zero. That *might* work.

That lovely kernel warning about "spidev" not being a shiny value in DT
is, well, just lovely. Yup, because patching kernel is definitely
something that one just wants to do. Oh well.

Change-Id: If9e444366fb4c4a2fc592ca99e84048a9c501da1