Loading different DTS files from U-Boot
We're at a point where it probably makes sense to separate the generic
"hey, let's experiment with this" Clearfog config from a device-specific
"I'm just a CL-ROADMv2" or "I'm a future SDN ROADM thingy".

The CL-ROADMv2 is branched out in order to cement this legacy, one-off
device in stone. Note that ADT7643 is still being loaded in a crazy way
as explained by https://gerrit.cesnet.cz/483 .

Change-Id: Ie86d37cf94c36077134afa54cacdae6962dd8c11