clearfog: Don't require J18 for CS GPIO
This means that we won't need to abuse the RST pin hogging for a reset
signal. On the other hand, we will need a way of resetting these
peripherals automatically. I hope to be able to use a few gates which
just fires upon several CS pins being active at the same time because
that looks like a reasonable condition -- once the code to initialize
all CSes is upstream.

Change-Id: I68b4bfb5a2b42050fcf1223d2f6eeaee96cc3970