Use KConfig for /cfg filesystem size
Because clearfog and Epia have vastly differing flash sizes, we will
have to specify the desired size of the /cfg filesystem template that we
are making. Let's convert this to a "package" so that we have easy
access to the Kconfig tools -- even though it really isn't about any
"package" per se.

Change-Id: Ic95a201ab7874711514bdc8b2e60b711db97ddc0