RAUC: turn this into a configurable target
I am not using the generic rootfs target because it is optimized for
using fakeroot, and because it extracts that tarball from rootfs-common.
Rather than fighting the hooks (ROOTFS_XXX_POST_GEN_HOOKS is the only
thing which runs outside of fakeroot) and doing extra work by unpacking,
let's go all the way down to raw Makefile.

Change-Id: I075204291fbeac8d1af97a0a7c1254054ef1f594