Upgrade and fix U-Boot
I want to have access to the Ethernet interfaces from within U-Boot for
setting of persistent MAC addresses. There were some relevant changes in
the latest master, and one also has to fix the Device Tree which still
assumes that the board is a Clearfog Pro while we are using a Clearfog
Base.

The script_addr_r -> scriptaddr change is a bit misterious for me -- it
seems that that variable was actually never part of upstream. It seems
that I made a mistake during one of the reabses of my clearfog patch
series...

On the bright side, this makes it possible to persistently store
Ethernet MAC addresses within the U-Boot environment:

  eth1addr=00:11:17:00:00:01
  eth2addr=00:11:17:00:00:02
  eth3addr=00:11:17:00:00:03

"eth1" is the left RJ-45, "eth2" is the right one, and "eth3" is the SFP
slot.

Change-Id: I887ffbf2e9c7fbab7baccd14b9b40ef42a566678