clearfog: LED indication of a boot progress
On our board, a reset activates several CS GPIOs at once. We have a gate
which triggers a reset of the MAX14830 and MCP23S08 chips on that
condition. We *could* somehow hack around this, but that just screams
"too fragile" to me (GPIO state on reboot != GPIO state on power on),
which is why I'm leaving this as-is.

So all LEDs will be off after a reboot -- fine. Once the kernel gets
around to setting up drivers for SPI chips, the green LED will start
blinking. This phase should be quite short because there's a HW watchdog
which will reboot the board again (and therefore turn off all LEDs)
unless the system manages to boot properly.

Once everything is up, we simply turn the green LED to a steady on
state.

Change-Id: I3fde56f95e0db92aa763e013d6e5567dc26d85ed