clearfog: fast-blink the white LED early on
Now that we're using an I2C LED driver with no HW reset at CPU-reboot,
the "we're OK" indicators remained on even after reboot.

Fix this by reseting the LED driver via its SW reset, and then start
rapid blinking the status LED in white color (with a reduced duty cycle
and brightness so that it's still reasonably pleasant).

I'm not doing this from the U-Boot script because this approach allows
me to get before that three-second countdown for autoboot.

Change-Id: I2c908a036308d19c8a1b79a607b1c63ec01e8ba6