clearfog: give up on I2C access to the PSUs
Their PMBus implementation requires PEC, and that one is broken because
the HW does not know that there is that LTC4316 address translator in
between the bus master and the PMBus devices. Without extra SW support
which does not exist yet, this is going to be the path of least
resistance.

Change-Id: Ie9b27626d9f6cc0483d78efbf5238ba1b28876d2