Delay SSH key copying after the sshd have finished generating them
Previously, we were getting this dependency "for free" because the
netopeer2-keystored package was blocking on sshd to start because it was
requsing sshd's keys for its own purposes.

Now that our keys are different (that change happened in the main
buildroot repo in our add-on branch), we no longer depended on sshd to
finish its key generation, and in turn this meant that the main sshd's
host keys were not being preserved in our /cfg:

  # ls -al /cfg/etc/ssh/
  total 8
  drwxr-xr-x    2 root     root          4096 Sep 12 18:37 .
  drwxr-xr-x    4 root     root          4096 Sep 13 18:39 ..
  -rw-------    1 root     root             0 Sep 12 18:37 ssh_host_rsa_key.NHsbsiTO7k

Fix this by waiting for sshd to finish starting up, and only copy the
SSH host keys afterwards.

Also split copying of OpenSSH's host keys from copying og
Netopeer2-server's host keys because these are conceptually unrelated.

Change-Id: Ic927ec2316595d1f0722e2aad097e0ae9ab60690