clearfog: bridge over all ethernet interfaces with STP enabled
This is close to what the OpenROADM MSA suggests (they use RSTP, and we
don't have the mstpd in place for that, yet).

I'm using an extra generator because I need to specify which MAC address
to use for the bridge (that's important for DHCP). I could probably
generate a drop-in or something, but hey, this big hammer works. The
reason for having to use a dynamic thing is that there's no
systemd-networkd feature to specify "please use MAC from the first
iface" - -so one has to read that data and write out a config file just
for that.

Change-Id: Ia12b53a93cbfe83c6c3e0472cd1910c4918e9abb