systemd: prepare the /etc overlay prior to launching systemd
systemd performs a bunch of critical tasks prior to loading any unit
files from disk, or running generators. Examples of this include the
/etc/machine-id setup (which we worked around by a bind mount some time
ago), or reading )and setting) the hostname configuration.

Rather than playing the whack-a-mole game over and over again by adding
just enough scripts to re-kick a given daemon/subsystem once a
particular /etc file becomes available, this change moves /etc setup
(which includes the overlay and the /cfg filesystem as well) into an
"initramfs" (without actually using a separate initramfs, of course).

Thanks to Mantas Mikulėnas (grawity on IRC) for suggesting this.

Change-Id: I9ec4d7f1243929ea498142d435ae6689c53b6097