clearfog: kernel: rework spidev handling for OCM
This is all about how to get spidev to export a SPI device to userspace.
We switched to a udev-style control last fall, but one upstream patch
which was supposed to provide auto-driver-bind in response to a udev
rule which sets the `driver_override` attribute has not been merged yet.
See [1] for reasons why it might not get merged, ever.

Adjust the userspace to bind a driver from within udev.

[1] https://www.spinics.net/lists/linux-spi/msg16384.html

Change-Id: I26f839ece7881baf56b8f0705623024e117a3a02