Sync and update the CI
Now that the logs, artifacts and essentially everything is in Swift, I'm
moving our mirror there as well. (The ironic part is that the Swift GW
is currently IPv4-only, but maybe that will change, and our VMs used to
be IPv6-only, and some of the original mirrors were IPv4-only as well...)

Here's how to seed the mirror in Swift:

 mkdir mirror
 cd mirror
 ln -s ~/work/cesnet/gerrit/github/buildroot/buildroot/dl buildroot
 cd ..
 openstack object create ci-artifacts-public mirror/buildroot/*/*tar* mirror/buildroot/*/*.zip

I wonder how we'll be dealing with cache purging, but hey, perhaps I can
claims that this is "for license compliance", yay.

Regarding the host-python and host-python3, here's how these failures
look like within the CI system:

 Python build finished successfully!
 The necessary bits to build these optional modules were not found:
 _bz2                  _curses               _curses_panel
 _dbm                  _gdbm                 _sqlite3
 _tkinter              _uuid                 nis
 readline
 To find the necessary bits, look in setup.py in detect_modules() for the module's name.

 The following modules found by detect_modules() in setup.py, have been
 built by the Makefile instead, as configured by the Setup files:
 _abc                  atexit                pwd
 time

My working speculation is that building dependencies of these two packages
up-front might solve this. Well, the hope never dies.

Change-Id: I14e57c0394a65e48761418abe9529bea8d1d8762