Sync and update the CI
Now that the logs, artifacts and essentially everything is in Swift, I'm
moving our mirror there as well. (The ironic part is that the Swift GW
is currently IPv4-only, but maybe that will change, and our VMs used to
be IPv6-only, and some of the original mirrors were IPv4-only as well...)

Here's how to seed the mirror in Swift:

 mkdir mirror
 cd mirror
 ln -s ~/work/cesnet/gerrit/github/buildroot/buildroot/dl buildroot
 cd ..
 openstack object create ci-artifacts-public mirror/buildroot/*/*tar* mirror/buildroot/*/*.zip

I wonder how we'll be dealing with cache purging, but hey, perhaps I can
claims that this is "for license compliance", yay.

There were some random build failures which I assumed to be related to
Python. They were not, it was a change in libyang which now requires an
out-of-tree build.

Change-Id: I14e57c0394a65e48761418abe9529bea8d1d8762