udev: Use a new way of udev-triggered program run
Our previous form, PROGRAM+=foo, causes a warning in the log, and the
spidev device does not get created:

  systemd-udevd[141]: /usr/lib/udev/rules.d/40-spi-lumentum-ocm.rules:2 PROGRAM key takes '==' or '!=' operator, assuming '==', but please fix it.

This sucks, obviously, but I'm in no mood of chasing when exactly
systemd upstream broke this feature.

The upside of this behavior change is that adjusting that
driver_override attribute via ATTR now works.

Change-Id: I52fa805dca8be060b627a5bbf70ddbdadec4cd3f