clearfog: optoswitch: export temperature to hwmon
The AD5593R chip that we're using as a DAC for VOA control (and
for the cross-bar switch control via GPIO) also integrates ADCs. Some of
them are used for monitoring the set voltage. There's also the final one
which provides a very coarse temperature indication. The datasheet says
that the resolution is about 3°C when averaged over five samples (...),
so it is probably only useful as a rough indication of short-circuits,
etc. But hey, let's make it available.

And of course it sucks because the Linux driver apparently returns
results which do not make sense. The raw value is OK, but the scaling
and offset are both way off.

Change-Id: I5ad4c92827bd06b167d90c03d3c32b3bc3a9f783