WIP: Refresh the BeagleBone Black image
Barebox won't build (it complains about <linux/list.h> not found and I
cannot easily solve that), so let's go back to U-Boot. The 2019.01
release is buggy in its recognizing of partition tables (`fatls`
complains, for example), and the whole thing reverts to a "legacy boot",
so let's use the latest release, the 2019.07.

WIP Missing items:
- dual-slot booting
- watchdog

Change-Id: I013f145cf2fe24e41ea0dafd225e3feb1854712c