Refresh the BeagleBone Black image
Barebox won't build (it complains about <linux/list.h> not found and I
cannot easily solve that), so let's go back to U-Boot.

The 2019.01 U-Boot release is buggy in its recognizing of partition
tables (`fatls` complains, for example), and the whole thing reverts to
a "legacy boot", so let's use the latest release, the 2019.07.

I'm also giving up on the g_cdc module because the board that I'm using
has a microUSB header for the debug console.

- FIXME: get dual-slot booting working
- FIXME: activate watchdog from U-Boot

Change-Id: I013f145cf2fe24e41ea0dafd225e3feb1854712c