BeagleBone Black: let the default FW "boot from the SD card"
I just don't want to hold that button at each power-up, and I have not
added support for flashing to eMMC yet. The default, preinstalled
bootlaoder tries to read uEnv.txt from the SDD card's boot partition and
execute `uenvcmd` from there, so let's make it happy and adapt
ourselves.

The default U-Boot installed on-device (2015.01-00001-gb2412df) doesn't
define ${scriptaddr}, so we have to provide it if we want our launchers
to succeed.

Change-Id: I50dc5177f4dd16435d4f5d17908db938e4cc5c3c