BeagleBone Black: Work around lack of U-Boot environment and SD card
I wanted to redirect the boot sequence to the SD card if it's plugged
in without pressing that magic button. This, however, means that we're
still running from within the old U-Boot located on the eMMC, and that
one does not have any environment support, apparently. This means that
one cannot really do a `setenv czechlight calibration`.

Let's hardcode the device type into the SD configuration for now. This
should be removed once we support flashing the eMMC.

While we're at it, initialize `/etc/machine-id` as well.

Change-Id: I02a3d607594e2c50b439f6415ce56bc4496785c7