journal: allow substantially more space for the log
...in an attempt to actually capture more useful data.

Also ensure that whatever gets printed to stdout/stderr, such as C++'s
own exception handling, is considered very important and dropped by
accident.

Also relax the rate limit a bit; at least that ITMA polling on the
coherent A/D box produces a ton of data, and the trace bucket for that
service gets throttled.

A `journalctl -f -u systemd-journald` running in parallel was very
useful in debugging this.

Change-Id: If18446c9ab6c928a16e4944f331279cef627a612
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/cla-sysrepo/+/1992
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/1992