CI: build images for BBB, too
This is going to burn twice as many CPU cycles (it's one more full job),
but at the same time it should ensure that there are no grave
regressions. I think it's manageable, and we might always hide it behind
a `check bbb` or something CI command in future.

Also, given that the standard build times are about 20 minutes within
the CI system, let's reduce the timw allotted for this task to a full
hour. That should still leave plenty of reserve.

Change-Id: I53034f0c7c39f791cf790cf2b3d8d2128e14f012