Calibration box service setup
...and also syncing of the dep tree.

The catch was that the previous versions hardcoded the `-clearfog`
suffix into the appliance name in the systemd unit template. However,
there was really no place with the same `clearfog` string was defined.
The closest one was the "compatible HW" identifier within RAUC which is
*exactly* what I would have wanted to use, but the catch is that the
clearfog builds used `czechlight-clearfog` in there (aaargh), and
there's just no way of changing it now as that would prevent me from
updating via RAUC. Meh, meh, meh.

Just changing the systemd unit and renaming the appliance within
cla-sysrepo look like a reasonable option, IMHO. We will have to change
this for future model which are HW-incompatible on the schematic side...

Change-Id: Ic447cb38e813aa3f3aba581c0d4d23a3621d6211
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/cla-sysrepo/+/2156
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/2156