bbb: use RAUC eMMC layout
This, unfortunately, means that the system will no longer boot from the
SD card:

- U-Boot is set to store its env in the p1 VFAT partition
- rootfs' /boot is something else (kernel, etc) than what gets mounted
on system's /boot
- rootfs' /boot/boot.scr contains device numbers which refer to the eMMC
block device, not to the SD card

I think it might be possible to add some workarounds, such as exporting
the number of an "active MMC device" to an U-Boot variable or something,
but hey, let's give up. The only thing that I need is *some* method of
flashing content to the device, and any distro is good enough for that.

Change-Id: Ia966448a05ed5ce52747f5e7169136232f5502c2